// job-portal-fe/src/model/experienceDto.ts
/**
 * job-portal
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: local
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { CompanyDto } from './companyDto';
import { DateRangeDto } from './dateRangeDto';
import { ReferenceDto } from './referenceDto';

export interface ExperienceDto { 
    company?: CompanyDto;
    dateRange?: DateRangeDto;
    id?: string;
    jobCategories?: Array<ReferenceDto>;
    occupation: string;
}