// job-portal-fe/src/model/applicationDto.ts
/**
 * job-portal
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: local
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { ReferenceDto } from './referenceDto';

export interface ApplicationDto { 
    applicant?: ReferenceDto;
    date?: Date;
    id?: string;
    jobPosition?: ReferenceDto;
    state?: ApplicationDto.StateEnum;
}
export namespace ApplicationDto {
    export type StateEnum = 'OPEN' | 'SEEN' | 'APPROVED' | 'DECLINED' | 'CLOSED';
    export const StateEnum = {
        OPEN: 'OPEN' as StateEnum,
        SEEN: 'SEEN' as StateEnum,
        APPROVED: 'APPROVED' as StateEnum,
        DECLINED: 'DECLINED' as StateEnum,
        CLOSED: 'CLOSED' as StateEnum
    };
}